add the fontstash file from the mio engine
This is an alternative implementation of a fontstash that I liked
quite a lot. Especially because in its current incarnation it uses
modern OpenGL. Of course fontstash can be made to do the same, it
is GL-agnostic. But this provides it all in one package and I'd
like to hack on it. On of the reason I do this is that my engine
is OpenGL only for simplicity's sake, and that's not going to change.

So there's no harm in having an OpenGL-only font stash. It also does
subpixel rendering, which would be quite easy to retrofit into vanilla
fontstash as well, but for now I don't want to make the effort to
change the hashing structure et cetera that would be necessary. It
probably won't get accepted into mainline anyway.

It was made by Tor Andersson.